+++ german_accounting/german_accounting/doctype/datev_action_panel/test_datev_action_panel.py
@@ -1,0 +1,8 @@
+# Copyright (c) 2024, phamos.eu and Contributors
+# See license.txt
+
+# import frappe
+import unittest
+
+class TestDATEVActionPanel(unittest.TestCase):
+	pass
