--- german_accounting/german_accounting/doctype/datev_export_mapping/test_datev_export_mapping.py
@@ -1,9 +1,0 @@
-# Copyright (c) 2024, phamos.eu and Contributors
-# See license.txt
-
-# import frappe
-from frappe.tests.utils import FrappeTestCase
-
-
-class TestDATEVExportMapping(FrappeTestCase):
-	pass
