--- german_accounting/german_accounting/doctype/datev_export_mapping_table/datev_export_mapping_table.py
@@ -1,8 +1,0 @@
-# Copyright (c) 2024, phamos.eu and contributors
-# For license information, please see license.txt
-
-# import frappe
-from frappe.model.document import Document
-
-class DATEVExportMappingTable(Document):
-	pass
