+++ german_accounting/german_accounting/workspace/german_accounting/german_accounting.json
@@ -1,0 +1,61 @@
+{
+ "category": "Domains",
+ "charts": [],
+ "creation": "2024-03-31 01:52:23.126747",
+ "developer_mode_only": 0,
+ "disable_user_customization": 0,
+ "docstatus": 0,
+ "doctype": "Workspace",
+ "extends_another_page": 0,
+ "hide_custom": 0,
+ "idx": 0,
+ "is_default": 0,
+ "is_standard": 1,
+ "label": "German Accounting",
+ "links": [
+  {
+   "hidden": 0,
+   "is_query_report": 0,
+   "label": "DATEV",
+   "link_to": "DATEV Export Log",
+   "link_type": "DocType",
+   "onboard": 0,
+   "type": "Card Break"
+  },
+  {
+   "hidden": 0,
+   "is_query_report": 0,
+   "label": "DATEV Action Panel",
+   "link_to": "DATEV Action Panel",
+   "link_type": "DocType",
+   "onboard": 0,
+   "type": "Link"
+  },
+  {
+   "hidden": 0,
+   "is_query_report": 0,
+   "label": "DATEV Export Logs",
+   "link_to": "DATEV Export Log",
+   "link_type": "DocType",
+   "onboard": 0,
+   "type": "Link"
+  },
+  {
+   "hidden": 0,
+   "is_query_report": 1,
+   "label": "DATEV Sales Invoice Export Report",
+   "link_to": "DATEV Sales Invoice Export",
+   "link_type": "Report",
+   "onboard": 0,
+   "type": "Link"
+  }
+ ],
+ "modified": "2024-04-02 18:30:25.687609",
+ "modified_by": "Administrator",
+ "module": "German Accounting",
+ "name": "German Accounting",
+ "owner": "Administrator",
+ "pin_to_bottom": 0,
+ "pin_to_top": 0,
+ "shortcuts": []
+}
