+++ german_accounting/patches.txt
@@ -1,1 +1,3 @@
 german_accounting.patches.v13_0.remove_vat_id_field_from_q_so_si
+german_accounting.patches.v13_0.remove_datev_mapping_related_doctypes
+german_accounting.patches.v13_0.remove_imat_workspace
