+++ german_accounting/patches/v13_0/remove_datev_mapping_related_doctypes.py
@@ -1,0 +1,12 @@
+# Copyright (c) 2019, Frappe and Contributors
+# License: GNU General Public License v3. See license.txt
+
+
+import frappe
+
+
+def execute():
+
+	if frappe.db.exists("DocType", "DATEV Export Mapping"):
+		frappe.delete_doc("DocType", "DATEV Export Mapping", force=1)
+		frappe.delete_doc("DocType", "DATEV Export Mapping Table", force=1)
