+++ german_accounting/patches/v13_0/remove_imat_workspace.py
@@ -1,0 +1,11 @@
+# Copyright (c) 2019, Frappe and Contributors
+# License: GNU General Public License v3. See license.txt
+
+
+import frappe
+
+
+def execute():
+
+	if frappe.db.exists("Workspace", "IMAT"):
+		frappe.delete_doc("Workspace", "IMAT", force=1)
